// flowforge/apps/desktop/src/litegraph/src/types/widgets.ts
export interface IWidgetOptions<TValues = unknown[]> {
  on?: string
  off?: string
  max?: number
  min?: number
  slider_color?: CanvasColour
  marker_color?: CanvasColour
  precision?: number
  read_only?: boolean
  /**
   * @deprecated Use {@link IWidgetOptions.step2} instead.
   * The legacy step is scaled up by 10x in the legacy frontend logic.
   */
  step?: number
  /** The step value for numeric widgets. */
  step2?: number

  y?: number
  multiline?: boolean
  // TODO: Confirm this
  property?: string
  /** If `true`, an input socket will not be created for this widget. */
  socketless?: boolean
  /** If `true`, the widget will not be rendered by the Vue renderer. */
  canvasOnly?: boolean
  /** Used as a temporary override for determining the asset type in vue mode*/
  nodeType?: string

  values?: TValues
  /** Optional function to format values for display (e.g., hash → human-readable name) */
  getOptionLabel?: (value?: string | null) => string
  callback?: IWidget['callback']
  iconClass?: string
}
// ...
export type IWidget =
  | IBooleanWidget
  | INumericWidget
  | IStringWidget
  | IComboWidget
  | IStringComboWidget
  | ICustomWidget
  | ISliderWidget
  | IButtonWidget
  | IKnobWidget
  | IFileUploadWidget
  | IColorWidget
  | IMarkdownWidget
  | IImageWidget
  | ITreeSelectWidget
  | IMultiSelectWidget
  | IChartWidget
  | IGalleriaWidget
  | IImageCompareWidget
  | ISelectButtonWidget
  | ITextareaWidget
  | IAssetWidget
  | IImageCropWidget
  | IBoundingBoxWidget
// ...
type ComboWidgetValues =
  | string[]
  | Record<string, string>
  | ((widget?: IComboWidget, node?: LGraphNode) => string[])

/** A combo-box widget (dropdown, select, etc) */
export interface IComboWidget extends IBaseWidget<
  string | number,
  'combo',
  RequiredProps<IWidgetOptions<ComboWidgetValues>, 'values'>
> {
  type: 'combo'
  value: string | number
}
// ...
export interface IBaseWidget<
  TValue = boolean | number | string | object | undefined,
  TType extends string = string,
  TOptions extends IWidgetOptions<unknown> = IWidgetOptions<unknown>
> {
  [symbol: symbol]: boolean

  linkedWidgets?: IBaseWidget[]

  name: string
  options: TOptions

  label?: string
  /** Widget type (see {@link TWidgetType}) */
  type: TType
  value?: TValue
  vueTrack?: () => void

  /**
   * Whether the widget value should be serialized on node serialization.
   * @default true
   */
  serialize?: boolean

  /**
   * The computed height of the widget. Used by customized node resize logic.
   * See scripts/domWidget.ts for more details.
   * @readonly [Computed] This property is computed by the node.
   */
  computedHeight?: number

  /**
   * The starting y position of the widget after layout.
   * @readonly [Computed] This property is computed by the node.
   */
  y: number

  /**
   * The y position of the widget after drawing (rendering).
   * @readonly [Computed] This property is computed by the node.
   * @deprecated There is no longer dynamic y adjustment on rendering anymore.
   * Use {@link IBaseWidget.y} instead.
   */
  last_y?: number

  width?: number
  /**
   * Whether the widget is disabled. Disabled widgets are rendered at half opacity.
   * See also {@link IBaseWidget.computedDisabled}.
   */
  disabled?: boolean

  /**
   * The disabled state used for rendering based on various conditions including
   * {@link IBaseWidget.disabled}.
   * @readonly [Computed] This property is computed by the node.
   */
  computedDisabled?: boolean

  hidden?: boolean
  advanced?: boolean
  /**
   * This property is automatically computed on graph change
   * and should not be changed.
   * Promoted widgets have a colored border
   * @see /core/graph/subgraph/proxyWidget.registerProxyWidgets
   */
  promoted?: boolean

  tooltip?: string

  // TODO: Confirm this format
  callback?(
    value: unknown,
    canvas?: LGraphCanvas,
    node?: LGraphNode,
    pos?: Point,
    e?: CanvasPointerEvent
  ): void

  /**
   * Simple callback for pointer events, allowing custom widgets to events relevant to them.
   * @param event The pointer event that triggered this callback
   * @param pointerOffset Offset of the pointer relative to {@link node.pos}
   * @param node The node this widget belongs to
   * @todo Expose CanvasPointer API to custom widgets
   */
  mouse?(
    event: CanvasPointerEvent,
    pointerOffset: Point,
    node: LGraphNode
  ): boolean
  /**
   * Draw the widget.
   * @param ctx The canvas context to draw on.
   * @param node The node this widget belongs to.
   * @param widget_width The width of the widget.
   * @param y The y position of the widget.
   * @param H The height of the widget.
   * @param lowQuality Whether to draw the widget in low quality.
   */
  draw?(
    ctx: CanvasRenderingContext2D,
    node: LGraphNode,
    widget_width: number,
    y: number,
    H: number,
    lowQuality?: boolean
  ): void

  /**
   * Compatibility method for widgets implementing the draw
   * method when displayed in non-canvas renderers.
   * Set by the current renderer implementation.
   * When called, performs a draw operation.
   */
  triggerDraw?: () => void

  /**
   * Compute the size of the widget. Overrides {@link IBaseWidget.computeSize}.
   * @param width The width of the widget.
   * @deprecated Use {@link IBaseWidget.computeLayoutSize} instead.
   * @returns The size of the widget.
   */
  computeSize?(width?: number): Size

  /**
   * Compute the layout size of the widget.
   * @param node The node this widget belongs to.
   * @returns The layout size of the widget.
   */
  computeLayoutSize?(
    this: IBaseWidget,
    node: LGraphNode
  ): {
    minHeight: number
    maxHeight?: number
    minWidth: number
    maxWidth?: number
  }

  /**
   * Callback for pointerdown events, allowing custom widgets to register callbacks to occur
   * for all {@link CanvasPointer} events.
   *
   * This callback is operated early in the pointerdown logic; actions that prevent it from firing are:
   * - `Ctrl + Drag` Multi-select
   * - `Alt + Click/Drag` Clone node
   * @param pointer The CanvasPointer handling this event
   * @param node The node this widget belongs to
   * @param canvas The LGraphCanvas where this event originated
   * @returns Returning `true` from this callback forces Litegraph to ignore the event and
   * not process it any further.
   */
  onPointerDown?(
    pointer: CanvasPointer,
    node: LGraphNode,
    canvas: LGraphCanvas
  ): boolean
}